requirements.txt
django
gunicorn
whitenoise
psycopg2-binary
requests
urllib3
django-cors-headers
djangorestframework
python-dotenv
